refactor: rename utils.ts to convert-api-track-to-player-track.ts and replace it to /src/player/utills
