+++ apps/rtk-query/src/pages/UserPage/ui/UserTabs/TracksTab/TracksTab.tsx
@@ -6,7 +6,7 @@ import { TrackActions } from '@/features/tracks/ui/TrackActions/TrackActions'
 import { TrackRow } from '@/features/tracks/ui/TrackRow/TrackRow'
 import { useOwnerData } from '@/pages/UserPage/hooks'
 import { selectCurrentPlaylistId, usePlayerControls, useQueueControls } from '@/player'
-import { convertApiTracksToPlayerTracks } from '@/player/utils.ts'
+import { convertApiTracksToPlayerTracks } from '@/player/utils/convert-api-track-to-player-track.ts'
 import noCoverPlaceholder from '@/shared/assets/images/no-cover-placeholder.avif'
 import { Button } from '@/shared/components'
 import { useAppSelector } from '@/shared/hooks'
